pyproject.toml
[project]
name = "question-generation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "bitsandbytes>=0.48.1",
    "google-genai>=1.46.0",
    "openai>=2.3.0",
    "optimum>=2.0.0",
    "pillow>=11.3.0",
    "polars>=1.34.0",
    "scikit-learn>=1.7.2",
    "torch>=2.8.0",
    "transformers>=4.57.0",
]

[tool.ruff]
# Always generate Python 3.11-compatible code.
target-version = "py311"
